Moved the open contacts list implementation in the Main Activity
